Python/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.2.0
blinker==1.7.0
boto3==1.35.9
botocore==1.35.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
fastapi==0.109.2
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
fortls==2.13.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.24
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PythonTurtle==0.3.2
redis==5.0.1
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.2
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
starlette==0.36.3
tldextract==5.1.1
typing_extensions==4.9.0
urllib3==2.2.2
uvicorn==0.27.0.post1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
